+++ src/Loader/Http/Browser/Screenshot.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Crwlr\Crawler\Loader\Http\Browser;
+
+class Screenshot
+{
+    public function __construct(
+        public readonly string $path,
+    ) {}
+}