+++ tests/Loader/Http/Messages/RespondedRequestTest.php
@@ -2,9 +2,12 @@
 
 namespace tests\Loader\Http\Messages;
 
+use Crwlr\Crawler\Loader\Http\Browser\Screenshot;
 use Crwlr\Crawler\Loader\Http\Messages\RespondedRequest;
 use GuzzleHttp\Psr7\Request;
 use GuzzleHttp\Psr7\Response;
+
+use function tests\helper_testfilesdir;
 
 it('can be created from request and response objects.', function () {
     $request = new Request('GET', '/');
@@ -138,6 +141,7 @@ it('can be serialized', function () {
     $respondedRequest = new RespondedRequest(
         new Request('POST', '/home', ['key' => 'val'], 'bod'),
         new Response(201, ['k' => 'v'], 'res'),
+        [new Screenshot('/path/to/screenshot.png'), new Screenshot('/another/path/to/screenshot.webp')],
     );
 
     $respondedRequest->addRedirectUri('/index');
@@ -145,14 +149,15 @@ it('can be serialized', function () {
     $serialized = serialize($respondedRequest);
 
     expect($serialized)->toBe(
-        'O:51:"Crwlr\Crawler\Loader\Http\Messages\RespondedRequest":8:{s:13:"requestMethod";s:4:"POST";s:10:' .
+        'O:51:"Crwlr\\Crawler\\Loader\\Http\\Messages\\RespondedRequest":9:{s:13:"requestMethod";s:4:"POST";s:10:' .
         '"requestUri";s:5:"/home";s:14:"requestHeaders";a:1:{s:3:"key";a:1:{i:0;s:3:"val";}}s:11:"requestBody";' .
         's:3:"bod";s:12:"effectiveUri";s:6:"/index";s:18:"responseStatusCode";i:201;s:15:"responseHeaders";a:1:{' .
-        's:1:"k";a:1:{i:0;s:1:"v";}}s:12:"responseBody";s:3:"res";}',
+        's:1:"k";a:1:{i:0;s:1:"v";}}s:12:"responseBody";s:3:"res";s:11:"screenshots";a:2:{i:0;' .
+        's:23:"/path/to/screenshot.png";i:1;s:32:"/another/path/to/screenshot.webp";}}',
     );
 });
 
-test('a serialized instance can be unserialized', function () {
+test('an old serialized instance without screenshots array can be unserialized', function () {
     $serialized = 'O:51:"Crwlr\Crawler\Loader\Http\Messages\RespondedRequest":8:{s:13:"requestMethod";s:4:"POST";' .
         's:10:"requestUri";s:5:"/home";s:14:"requestHeaders";a:1:{s:3:"key";a:1:{i:0;s:3:"val";}}s:11:"requestBody";' .
         's:3:"bod";s:12:"effectiveUri";s:6:"/index";s:18:"responseStatusCode";i:201;s:15:"responseHeaders";a:1:{' .
@@ -173,7 +178,37 @@ test('a serialized instance can be unserialized', function () {
         ->and($respondedRequest->response->getBody()->getContents())->toBe('res');
 });
 
-it('can be created from a serialized array', function () {
+test('a serialized instance can be unserialized', function () {
+    // We need actual existing file paths for screenshots
+    $screenshot1 = helper_testfilesdir('screenshot1.png');
+
+    $screenshot2 = helper_testfilesdir('screenshot2.jpeg');
+
+    $serialized = 'O:51:"Crwlr\Crawler\Loader\Http\Messages\RespondedRequest":9:{s:13:"requestMethod";s:4:"POST";' .
+        's:10:"requestUri";s:5:"/home";s:14:"requestHeaders";a:1:{s:3:"key";a:1:{i:0;s:3:"val";}}s:11:"requestBody";' .
+        's:3:"bod";s:12:"effectiveUri";s:6:"/index";s:18:"responseStatusCode";i:201;s:15:"responseHeaders";a:1:{' .
+        's:1:"k";a:1:{i:0;s:1:"v";}}s:12:"responseBody";s:3:"res";s:11:"screenshots";a:2:{i:0;' .
+        's:' . strlen($screenshot1) . ':"' . $screenshot1 . '";i:1;' .
+        's:' . strlen($screenshot2) . ':"' . $screenshot2 . '";}}';
+
+    $respondedRequest = unserialize($serialized);
+
+    /** @var RespondedRequest $respondedRequest */
+
+    expect($respondedRequest)->toBeInstanceOf(RespondedRequest::class)
+        ->and($respondedRequest->request->getMethod())->toBe('POST')
+        ->and($respondedRequest->request->getUri()->__toString())->toBe('/home')
+        ->and($respondedRequest->request->getHeaders())->toBe(['key' => ['val']])
+        ->and($respondedRequest->request->getBody()->getContents())->toBe('bod')
+        ->and($respondedRequest->effectiveUri())->toBe('/index')
+        ->and($respondedRequest->response->getStatusCode())->toBe(201)
+        ->and($respondedRequest->response->getHeaders())->toBe(['k' => ['v']])
+        ->and($respondedRequest->response->getBody()->getContents())->toBe('res')
+        ->and($respondedRequest->screenshots[0]->path)->toBe($screenshot1)
+        ->and($respondedRequest->screenshots[1]->path)->toBe($screenshot2);
+});
+
+it('can be created from an old serialized array that was not containing the screenshots array', function () {
     $serialized = 'a:8:{s:13:"requestMethod";s:3:"GET";s:10:"requestUri";s:4:"/foo";s:14:"requestHeaders";a:0:{}s:11:' .
         '"requestBody";s:0:"";s:12:"effectiveUri";s:4:"/bar";s:18:"responseStatusCode";i:200;s:15:"responseHeaders";' .
         'a:0:{}s:12:"responseBody";s:0:"";}';
@@ -185,10 +220,50 @@ it('can be created from a serialized array', function () {
         ->and($respondedRequest->effectiveUri())->toBe('/bar');
 });
 
+it('can be created from a serialized array that is containing the screenshots array', function () {
+    // We need actual existing file paths
+    $screenshot1 = helper_testfilesdir('screenshot1.png');
+
+    $screenshot2 = helper_testfilesdir('screenshot2.jpeg');
+
+    $serialized = 'a:9:{s:13:"requestMethod";s:3:"GET";s:10:"requestUri";s:4:"/foo";s:14:"requestHeaders";a:0:{}s:11:' .
+        '"requestBody";s:0:"";s:12:"effectiveUri";s:4:"/bar";s:18:"responseStatusCode";i:200;s:15:"responseHeaders";' .
+        'a:0:{}s:12:"responseBody";s:0:"";s:11:"screenshots";a:2:{i:0;' .
+        's:' . strlen($screenshot1) . ':"' . $screenshot1 . '";i:1;' .
+        's:' . strlen($screenshot2) . ':"' . $screenshot2 . '";}}';
+
+    $respondedRequest = RespondedRequest::fromArray(unserialize($serialized));
+
+    expect($respondedRequest)->toBeInstanceOf(RespondedRequest::class)
+        ->and($respondedRequest->request->getUri()->__toString())->toBe('/foo')
+        ->and($respondedRequest->effectiveUri())->toBe('/bar')
+        ->and($respondedRequest->screenshots[0]->path)->toBe($screenshot1)
+        ->and($respondedRequest->screenshots[1]->path)->toBe($screenshot2);
+});
+
+test(
+    'when creating from a serialized array, it checks screenshot paths for existence and throws away screenshots ' .
+    'when the files don\'t exist',
+    function () {
+        $serialized = 'a:9:{s:13:"requestMethod";s:3:"GET";s:10:"requestUri";s:4:"/foo";s:14:"requestHeaders";' .
+            'a:0:{}s:11:"requestBody";s:0:"";s:12:"effectiveUri";s:4:"/bar";s:18:"responseStatusCode";i:200;' .
+            's:15:"responseHeaders";a:0:{}s:12:"responseBody";s:0:"";s:11:"screenshots";a:2:{i:0;' .
+            's:24:"/path/to/screenshot1.png";i:1;s:25:"/path/to/screenshot2.jpeg";}}';
+
+        $respondedRequest = RespondedRequest::fromArray(unserialize($serialized));
+
+        expect($respondedRequest)->toBeInstanceOf(RespondedRequest::class)
+            ->and($respondedRequest->request->getUri()->__toString())->toBe('/foo')
+            ->and($respondedRequest->effectiveUri())->toBe('/bar')
+            ->and($respondedRequest->screenshots)->toHaveCount(0);
+    },
+);
+
 it('has a toArrayForResult() method', function () {
     $respondedRequest = new RespondedRequest(
         new Request('POST', '/home', ['key' => 'val'], 'bod'),
         new Response(201, ['k' => 'v'], 'res'),
+        [new Screenshot('/path/to/screenshot.jpg')],
     );
 
     expect($respondedRequest->toArrayForResult())->toBe([
@@ -200,6 +275,7 @@ it('has a toArrayForResult() method', function () {
         'responseStatusCode' => 201,
         'responseHeaders' => ['k' => ['v']],
         'responseBody' => 'res',
+        'screenshots' => ['/path/to/screenshot.jpg'],
         'url' => '/home',
         'uri' => '/home',
         'status' => 201,